refactor(Calculator): split Calculator to model and controller
