+++ src/main/java/calculator/Application.java
@@ -1,13 +1,15 @@
 package calculator;
 
-import calculator.controller.Calculator;
+import calculator.controller.CalculatorController;
 import calculator.controller.io.ConsoleInputHandler;
 import calculator.controller.io.ConsoleOutputHandler;
+import calculator.model.CalculatorModel;
 
 public class Application {
     public static void main(String[] args) {
         // TODO: 프로그램 구현
-        Calculator calculator = new Calculator(new ConsoleInputHandler(), new ConsoleOutputHandler());
+        CalculatorController calculator = new CalculatorController(new ConsoleInputHandler(),
+                new ConsoleOutputHandler(), new CalculatorModel());
         calculator.run();
     }
 }
