+++ src/main/java/calculator/controller/io/OutputHandler.java
@@ -1,4 +1,4 @@
-package calculator.controller;
+package calculator.controller.io;
 
 public interface OutputHandler {
     void displayResult(Integer result);
